==> Step 1 - cleaning survey data.ipynb <==
import pandas as pd
import numpy as np

#load data
df_survey = pd.read_csv('Data for thesis students/data.csv')
df_survey = df_survey.drop(df_survey.columns[0], axis = 1)

#df_survey.head()

#only keep relevant columns
df_demographics = df_survey[['EthicaID', 'Age', 'Sex']] 

#only keep one row of information for each participant
df_demographics = df_demographics.drop_duplicates(subset = ['EthicaID']) 

#df_demographics.head()

#check for missing values
#print(df_demographics.isna().sum()) #no missing values

#print (sorted(df_demographics['Age'].unique())) #one illegal value -0.4166667

df_demographics = df_demographics.replace(-0.41666670000000006, np.nan) #make illeagle vlaue to NaN
#print(df_demographics.isna().sum()) #check if it is really NaN

#df_demographics.head()

#save cleanded dataframe to csv for later use
df_demographics.to_csv('cleaned_data/demographics.csv')
