==> Step 2 - cleaning playstore.ipynb <==
import pandas as pd

playstore = pd.read_csv('Additional data/Google-Playstore.csv')

playstore.head()

playstore['Category'] = playstore['Category'].replace(['Adventure', 'Racing', 'Puzzle', 'Arcade', 'Card', 
                                                       'Action', 'Simulation','Role Playing', 'Casino', 
                                                       'Libraries & Demo', 'Casual', 'Trivia', 'Board', 
                                                       'Educational', 'Strategy', 'Word', 'Music'], 'Games')

playstore['Category'].unique()

len(playstore['Category'].unique())

playstore_short = playstore[['App Id', 'App Name', 'Category']]

playstore_short.to_csv('cleaned_data/playstore.csv')

categories = pd.read_csv('Additional data/app_categories.csv')

categories.head()

categories_short = categories[['app_id', 'name', 'category', 'better_category']]

categories_short['better_category'].unique()

categories_short.to_csv('cleaned_data/categories_short.csv')
